Introduce package api and move endpoints routes there
