+++ go.mod
@@ -1,14 +1,16 @@
 module github.com/ppacer/core
 
-go 1.21
+go 1.22
 
-require modernc.org/sqlite v1.25.0
+require (
+	github.com/lib/pq v1.10.9
+	modernc.org/sqlite v1.25.0
+)
 
 require (
 	github.com/dustin/go-humanize v1.0.1 // indirect
 	github.com/google/uuid v1.3.0 // indirect
 	github.com/kballard/go-shellquote v0.0.0-20180428030007-95032a82bc51 // indirect
-	github.com/lib/pq v1.10.9 // indirect
 	github.com/mattn/go-isatty v0.0.19 // indirect
 	github.com/remyoudompheng/bigfft v0.0.0-20230129092748-24d4a6f8daec // indirect
 	golang.org/x/mod v0.12.0 // indirect
